% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_og_card.R
\name{add_og_card}
\alias{add_og_card}
\title{Create open graph meta tags}
\usage{
add_og_card(card_type = c("website", "article", "book"), title, image,
  url, description = NULL, image_alt = NULL, site_name = NULL,
  video = NULL, audio = NULL, author = NULL, file = "og_card.html")
}
\arguments{
\item{card_type}{Required. Open graph card type. See http://ogp.me/#types for list}

\item{title}{Required. Title for card content.}

\item{image}{Required. Link for the content image.}

\item{url}{Required. Canonical URl to be used as permanent ID in the graph.}

\item{description}{Description for card content. One to two sentences.}

\item{image_alt}{Alternate text to display if image fails.}

\item{site_name}{If your object is part of a larger web site, the name which should be displayed for the overall site.}

\item{video}{URL to a video file that complements the object}

\item{audio}{URL to an audio file that complements the object}

\item{author}{Authors of the content if content type is "article" or "book"}

\item{file}{Required. File name to put in the header of Rmd doc, must be .html.}
}
\description{
An easy way to create open graph meta tags for your blogdown post.
    For more details on the meta tags see http://ogp.me/
}
\examples{
add_og_card(title = "HeadR is great!",
       image = "https://www.rstudio.com/wp-content/uploads/2014/06/RStudio-Ball.png",
       url = "https://www.jumpingrivers.com",
       file = "hello.html")
}
