% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_twitter_card.R
\name{add_twitter_card}
\alias{add_twitter_card}
\title{Create twitter summary cards}
\usage{
add_twitter_card(card_type = "summary", user = "@jumping_uk",
  content_title = "Content title",
  content_description = "Content description.",
  content_image = "https://www.rstudio.com/wp-content/uploads/2014/06/RStudio-Ball.png",
  file = "twitter-card.html")
}
\arguments{
\item{card_type}{must be "summary" or "summary_large_image"}

\item{user}{twitter user id for the card to be linked to}

\item{content_title}{title for card content, max 70 characters}

\item{content_description}{description for card content, max 200 characters}

\item{content_image}{link for the content image}

\item{file}{name to put in the header of Rmd doc}
}
\description{
An easy way to create twitter summary cards for your blogdown post
}
\examples{
add_twitter_card()
}
