% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_og_card.R
\name{check_og_arguments}
\alias{check_og_arguments}
\title{Checking function arguments}
\usage{
check_og_arguments(card_type, title, image, url, file)
}
\arguments{
\item{card_type}{Required. Open graph card type. See http://ogp.me/#types for list}

\item{title}{Required. Title for card content.}

\item{image}{Required. Link for the content image.}

\item{url}{Required. Canonical URl to be used as permanent ID in the graph.}

\item{file}{Required. File name to put in the header of Rmd doc, must be .html.}
}
\description{
Checking function arguments
}
