% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_twitter_card.R
\name{check_twitter_arguments}
\alias{check_twitter_arguments}
\title{checking function arguments}
\usage{
check_twitter_arguments(card_type, title, description, file)
}
\arguments{
\item{card_type}{Required. Twitter card type. Must be "summary" or "summary_large_image".}

\item{title}{Required. Title for card content. Max 70 characters.}

\item{description}{Description for card content, max 200 characters.}

\item{file}{Required. File name to put in the header of Rmd doc, must be .html.}
}
\description{
checking function arguments
}
