% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cipoly.R
\name{cipoly}
\alias{cipoly}
\title{Do an area shaded confidence interval with transparency}
\usage{
cipoly(x, b, se, t = 1.96, col.fill = "darkgrey", alpha.fill = 0.4)
}
\arguments{
\item{x}{the x-axis variable in the plot, often time}

\item{b}{coefficients or means}

\item{se}{standard errors}

\item{t}{number of standard errors determining width}
}
\description{

}
\details{
A version of ribbon where user provides estimates and se
}
