% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dir.R
\name{clean_dir}
\alias{clean_dir}
\title{Remove leftover LaTeX files}
\usage{
clean_dir(
  path = ".",
  ext = c("aux", "snm", "toc", "log", "out", "nav", "gz", "lof", "lot"),
  del = FALSE,
  bib = TRUE
)
}
\arguments{
\item{path}{folder, defaults to current directory}

\item{ext}{character vector of file extensions, defaults to 7 Beamer extensions (not tex or pdf!) and two book extensions}

\item{del}{logical to say whether to delete or just list the files, defaults to FALSE}

\item{bib}{logical to say whether to delete bbl and blg files, defaults to TRUE}
}
\description{
Remove leftover LaTeX files
}
\details{
defaults to concatenating all the files Beamer creates, but can change to get list of other sets of files
when del= TRUE the files are removed.
#' @examples
clean_dir("Lectures/Slides2018")
clean_dir("Lectures/Slides2018",del=TRUE)
clean_dir()
}
