% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct.R
\name{distinct}
\alias{distinct}
\title{Count number of distinct elements in a vector}
\usage{
distinct(x)
}
\arguments{
\item{x}{a numeric or character vector}
}
\description{
Count number of distinct elements in a vector
}
