% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{duplicates}
\alias{duplicates}
\title{Count and tag duplicates (in terms of a variable list)}
\usage{
duplicates(DT, by.vars, tag = FALSE)
}
\arguments{
\item{DT}{a data.table}

\item{by.vars}{a list of variables}

\item{tag}{should the fn create an indicator for duplicate obs?}
}
\description{

}
\details{
duplicates resembles Stata command of same name to count and tag repeat observations
}
