% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains.R
\name{\%contains\%}
\alias{\%contains\%}
\title{Operator to test if a string is present in another string}
\usage{
x \%contains\% y
}
\arguments{
\item{x}{a string}

\item{y}{a string}
}
\description{
Operator to test if a string is present in another string
}
\details{
Returns TRUE if the variable to left has the string to the right. Going forward use \%like\% instead.
}
