% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excludes.R
\name{\%excludes\%}
\alias{\%excludes\%}
\title{Operator to see if a string does not contain another string}
\usage{
x \%excludes\% y
}
\arguments{
\item{x}{a character vector}

\item{y}{a character vector}
}
\description{
Operator to see if a string does not contain another string
}
\details{
Returns a logical if LHS does not have RHS inside it
}
