% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ni.R
\name{\%ni\%}
\alias{\%ni\%}
\title{Not-in operator}
\usage{
x \%ni\% y
}
\arguments{
\item{x}{a vector, usually a set of id strings}

\item{y}{a vector, usually a set of id strings}
}
\description{
Negates the \code{\%in\%} operator to exclude elements of a set.
}
\details{
Returns a logical vector. Equivalent to \code{x != y[1] & x != y[2]} etc.,
but note that \code{y[1]} here is just shorthand -- the actual negation is
computed via \code{!\%in\%} and works for any length \code{y}.
}
\examples{
c(1, 5, 7, -99, NA) \%ni\% c(-99, NA)
all <- c("Coca-Cola", "Others", "Pepsi", "Private Label")
brands <- all[all \%ni\% c("Others", "Private Label")]
}
