% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_str.R
\name{\%+\%}
\alias{\%+\%}
\title{Paste together strings}
\usage{
x \%+\% y
}
\arguments{
\item{x}{a vector, usually string}

\item{y}{a vector, usually string}
}
\description{
#' @description paste two strings into one with no space between them
}
\details{
a plus operator to do a simple paste0, similar to Stata
}
\examples{
n = 5
"filename" \%+\% n \%+\% ".pdf"
c(1,NA,3,NA) \%+\% c(7,5,NA,NA)
}
