% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_na.R
\name{\%+na\%}
\alias{\%+na\%}
\title{when adding an NA keep the non-missing term in the sum}
\usage{
x \%+na\% y
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}
}
\description{
when adding an NA keep the non-missing term in the sum
}
\details{
sum that does not fail when adding a missing and a non-missing:
used in HeadR::merge_stata
#' @examples
2 \%+na\% NA
c(1,NA,3,NA) \%+na\% c(7,5,NA,NA)
}
