% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract.r
\name{hdr_abstract}
\alias{hdr_abstract}
\title{hdr_abstract}
\usage{
hdr_abstract(meta = yaml::yaml.load_file("_metadata.yaml"))
}
\arguments{
\item{meta}{metadata content}
}
\value{
string
}
\description{
Format the abstract
}
\details{
Read the metadata and inject the formatted abstract in a md document
}
