% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.r
\name{hdr_metadata}
\alias{hdr_metadata}
\title{hdr_metadata}
\usage{
hdr_metadata(meta = yaml::yaml.load_file("_metadata.yaml"),
  template = tpl_aps())
}
\arguments{
\item{meta}{metadata content}

\item{template}{journal-specific template}
}
\value{
string
}
\description{
Format the metadata
}
\details{
Read the metadata and inject the formatted string in a md document
}
