% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipolate.R
\name{ipolate}
\alias{ipolate}
\title{Linear interpolation of missing data}
\usage{
ipolate(xvar, yvar, extrapolate = FALSE)
}
\arguments{
\item{xvar}{the variable to interpolate using (usually year)}

\item{yvar}{the name of the variable with missing data}
}
\description{
Linear interpolation of missing data
}
\details{
ipolate resembles Stata command of same name to linearly interpolate missing data, and carry forward last observation or backward first observation if extrapolate=TRUE.
If you want the identical behavior to Stata, use ipolate_stata() instead.
}
\examples{
# examples
years <- 2000:2005
values <- c(1, NA, 3, NA, 5, NA)
ipolate(years, values) # should return c(1,2,3,4,5,NA)
ipolate(years, values, extrapolate=TRUE) # should return c(1,2,3,4,5,5)

years <- 2000:2005
values <- c(NA, 2, 3, NA, 5, NA)
ipolate(years, values) # should return c(NA,2,3,4,5,NA)
ipolate(years, values, extrapolate=TRUE) # should return c(2,2,3,4,5,5)

}
