% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_replace.R
\name{match_replace}
\alias{match_replace}
\title{Match codes from something to something}
\usage{
match_replace(code, from, to, dictionary)
}
\arguments{
\item{code}{code to map}

\item{from}{name of column from which to map}

\item{to}{name of column to which to map}

\item{dictionary}{data.frame with mapping}
}
\description{
Match codes from something to something
}
