% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_stata.R
\name{merge_stata}
\alias{merge_stata}
\title{merge.data.table, all, with Stata _merge tracking}
\usage{
merge_stata(DTx, DTy, by)
}
\arguments{
\item{DTx}{the master data.table}

\item{DTy}{the using data.table}

\item{by}{the variables to merge by as a string vector}
}
\description{

}
\examples{
DT1 <- data.table(i=1:5,x1 = c(1,3,2,4,8))
DT2 <- data.table(i=c(3,4,6),y1 = c(7,5,7))
DT3 <- merge_stata(DT1,DT2,"i")
}
