% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice.R
\name{nice}
\alias{nice}
\title{Make nice numbers}
\usage{
nice(x, digits = 2)
}
\arguments{
\item{x}{a number}

\item{digits}{the number of digits to show to right of decimal point}
}
\description{
Format numbers with a fixed number of decimals for a table

Warning the result is not numeric anymore, no more calculations
}
\examples{
 nice(c(17.0003,4.726677))
nice(c(17.0003,4.726677),digits=0)
}
