% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribbon.R
\name{ribbon}
\alias{ribbon}
\title{Do an area shaded ribbon with transparency}
\usage{
ribbon(
  x,
  y_lo,
  y_hi,
  col.fill = "darkgrey",
  alpha.fill = 0.4,
  col.edge = "darkgrey",
  alpha.edge = 0.4
)
}
\arguments{
\item{x}{the x-axis variable in the plot, often time}

\item{y_lo}{lower edge of the ribbon}

\item{y_hi}{upper edge of the ribbon}

\item{alpha.fill}{transparency parameter, lower is lighter}

\item{alpha.edge}{transparency parameter, set to 1 for no transparency}

\item{coll.fill}{colour to fill the ribbon}

\item{coll.edge}{colour to outline the upper and lower bounds}
}
\description{
Do an area shaded ribbon with transparency
}
\details{
A more general version of cipoly, for example when you have bootstrap CIs
}
\examples{
plot(1:10,type="n",ylim=c(0,15))
ribbon(1:10,2:11,4:13,col.fill="blue",col.edge="blue",alpha.edge=1)
ribbon(1:10,11:2,13:4,col.fill="red",col.edge="red")
}
