% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{scatter}
\alias{scatter}
\title{Wrapper for plot/points that defaults to filled transparent circles}
\usage{
scatter(..., col = "black", a = 0.5, pch = 16, add = FALSE)
}
\arguments{
\item{col}{colour, defaults to black}

\item{a}{abbreviation for alpha.f, 0=perfectly transparent, 1=solid}

\item{pch}{point type, defaults to filled circles}
}
\description{
Wrapper for plot/points that defaults to filled transparent circles
}
\details{
To avoid all the typing to get transparent filled circles which is what i want 99\% of the time for a scatter plot, i have created this wrapper. Since it defaults to using the same function one should be able to use it in mapply or lapply
}
