% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_words.R
\name{shared_words}
\alias{shared_words}
\title{Count (or obtain fraction of) shared words in a sentence}
\usage{
shared_words(
  x,
  y,
  sep = " ",
  rm = c("punct", "digit"),
  frac = FALSE,
  common = NULL
)
}
\arguments{
\item{x}{character vector}

\item{y}{character vector}

\item{sep}{character such as "_", defaults to " "}

\item{rm}{character vector consisting of "punct" and/or "digit"}

\item{frac}{logical, if TRUE function returns fraction of common words, with the longer sentence length as the denominator}

\item{common}{character vector of words that are so common they should not be used}
}
\description{
Count (or obtain fraction of) shared words in a sentence
}
\details{
The typical use of this function is to determine whether two vectors of names are close to each other.
The function assumes that each element of each vector is a sentence with " " or some other separator,
I recommend sep = "_" which is what my st_standard creates.
Warning: same sets of common words still have frac of shared=1, but count of shared = 0
}
\examples{
shared_words(c("HEFEI","I KNOW","GAS","GAS"),c("HEFEI II","I KNOW","GAS-LIGHT","GAS LIGHT"))
shared_words(c("HEFEI","I_KNOW","NO"),c("HEFEI_II","I_KNOW","GAS_LIGHT"),sep = "_",rm="digit",frac=TRUE)
cw <- c("NORTH","SOUTH")
shared_words(c("HEFEI","I_KNOW","NORTH_DAKOTA"),c("HEFEI_II","I_KNOW","NORTH_CAROLINA"),sep = "_",rm="digit",frac=TRUE,common=cw)
}
