% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_all.R
\name{st_all}
\alias{st_all}
\title{Take a vector of strings and tack sorted uniques together separating with "-"}
\usage{
st_all(x, sep = "-")
}
\arguments{
\item{x}{a numeric or string  vector}
}
\description{

}
\details{
returns a single string from a vector, useful in by commands to keep all the unique values, sorted and hyphenated
Default separator not work well with negative numbers or if string already contains hyphenation, switch to "|"
}
\examples{
st_all(c(1,3,1,2))
st_all(c("AZ","Pfizer","Moderna","AZ"))
st_all(c(1,3,-1,2),"|")
}
