% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_left.R
\name{st_left}
\alias{st_left}
\title{Create a new string of text to the left of a "_"}
\usage{
st_left(x, sep = "_")
}
\arguments{
\item{x}{a character vector}
}
\description{

}
\details{
use to separate a string into just the part to left of a delimiter such as "-" or "_", also see st_right
}
