% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_range.R
\name{st_range}
\alias{st_range}
\title{Calculate and display range of a vector of numbers or strings}
\usage{
st_range(x)
}
\arguments{
\item{x}{a numeric or string  vector}
}
\description{

}
\details{
returns a character with the lowest and highest values separated by "--" (as in LaTeX)
}
\examples{
st_range(c(1,3,-1,2))
st_range(c("az","zc","x","b"))
}
