% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_right.R
\name{st_right}
\alias{st_right}
\title{Create a new string of text to the right of a "_"}
\usage{
st_right(x, sep = "_")
}
\arguments{
\item{x}{a character vector}

\item{sep}{a separator character (defaults to "_")}
}
\description{

}
\details{
use to separate a string into just the part to right of a delimiter such as "-" or "_", also see st_left
}
