% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_standard.R
\name{st_standard}
\alias{st_standard}
\title{Standardize a name following set of procedures}
\usage{
st_standard(x)
}
\arguments{
\item{x}{string vector}
}
\description{

}
\details{
st_standard capitalizes and trims leading and following white space
It replaces a single "-" and multiple internal spaces between words with a single "_"
It also eliminates all other punctuation
The current version does not replace accented characters but i would like to do so later.
}
\examples{
st_standard(c("General Pacheko #2","B.M.W.","Fiat-Chrysler","Fiat Chrysler","Fiat chrysler","Austin, TX (USA)"))
}
