% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str2tex.R
\name{str2tex}
\alias{str2tex}
\title{create a column in a data.table with tabular info}
\usage{
str2tex(x, digits = 2)
}
\arguments{
\item{x}{a list of variables}

\item{digits}{number of digits after decimal point}
}
\description{
create a column in a data.table with tabular info
}
\details{
str2tex() is designed to be used inside a data.table. it takes a list of variables and checks for which are whole numbers.
the whole numbers are not altered. the real numbers are formatted fixed. It is the updated version of texout() function.
LaTeX column separators "&" are place between numbers and  backslashes are used at the end.
The last line is without "\\". Also keep the Int format of last element (if applicable)
}
\examples{
mydt <- data.table(ida=LETTERS[1:10],idn=1:10,x=runif(10),y=runif(10))
mydt[,output := str2tex(list(ida,idn,x,y))]
writeLines(mydt$output)
mydt[,output := str2tex(list(ida,idn,x,y),3)]
writeLines(mydt$output)
}
