% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str2tex_col.R
\name{str2tex_col}
\alias{str2tex_col}
\title{create a column in a data.table with tabular info}
\usage{
str2tex_col(data, unirounding = NULL, rounding = NULL, col_select = NULL)
}
\arguments{
\item{data}{a list of variables or a data.table}

\item{unirounding}{number of digits after decimal point}

\item{rounding}{a list of number of digits after decimal point}

\item{col_select}{a list of column numbers to select}
}
\description{
create a column in a data.table with tabular info
}
\details{
str2tex_col() is designed to be used inside a data.table. it takes a list of variables and a list of round digits with selected columns.
the whole numbers are altered as well. the real numbers are formatted fixed. It is the updated version of texout() function.
LaTeX column separators "&" are place between numbers and  backslashes are used at the end.
}
\examples{
data <- data.table(A = 1:5,B = c(1, 2.1234, 3.1234, 4.1234, 5.1234),C = c(1000.55, 2000.75, 3000.95, 4000.15, 5000.35),D = 1:5,E = c("a", "b", "c", "d", "e"))
data[,output := str2tex_col(list(A,B,C,D), rounding = c(0, 2, 1, 0))]
writeLines(data$output)
# apply the same rounding to all columns
data[,output := str2tex_col(list(A,B,C,D,E), unirounding = 3)]
writeLines(data$output)
# alternatively, input a data.table and select columns
formatted_output <- str2tex_col(data, rounding = c(0, 2, 1, 0), col_select = c(1, 2, 3, 4))
print(formatted_output)
}
