% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strpos.R
\name{strpos}
\alias{strpos}
\title{Find position in a string where the first instance of a pattern begins}
\usage{
strpos(chr, pat)
}
\arguments{
\item{chr}{string vector}

\item{pat}{pattern to locate}
}
\description{
Find position in a string where the first instance of a pattern begins
}
\details{
strpos resembles Stata command of same name, returns 0 if pat is absent from chr.
This new version is vectorized (the old one failed on the 3rd example)
}
\examples{
strpos("Coca-Cola","-")
strpos("Coca-Cola","Z")
strpos(c("WUHU_1","WU_2","WU_HU_WU","WUHUHU_3"),"_")
}
