% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where_missing.R
\name{where_missing}
\alias{where_missing}
\title{List of variables in a data.table with the count of missings}
\usage{
where_missing(DT)
}
\arguments{
\item{DT}{a data.table}
}
\description{
List of variables in a data.table with the count of missings
}
\details{
where_missing returns a reshaped data table filter to show just variables with missings along with the number of them. if the data.table has no missings, an Empty data.table is returned.
You must library(HeadR) to use this function
}
