% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_twitter_card.R
\name{add_twitter_card}
\alias{add_twitter_card}
\title{Create twitter summary cards}
\usage{
add_twitter_card(card = "summary", user = "@Jumping_UK",
  content_title = "max 70 characters",
  content_description = "max 200 characters",
  content_image = "https://www.google.co.uk/url?sa=i&rct=j&q=&esrc=s&source=images&cd=&ved=2ahUKEwiChLjtipfdAhVOxxoKHWfTD_oQjRx6BAgBEAU&url=https\%3A\%2F\%2Fwww.jumpingrivers.com\%2Fblog\%2Four-logo-in-r\%2F&psig=AOvVaw15HKIqgjXe2Hs8iFvu98k4&ust=1535797811922853",
  file = "foo.html")
}
\arguments{
\item{card}{must be "summary"}

\item{user}{twitter suer id for the card to be linked to}

\item{content_title}{title for card content, max 70 characters}

\item{content_description}{description for card content, max 200 characters}

\item{content_image}{link for the content image}

\item{file}{name to put in the header of Rmd doc}
}
\description{
An easy way to create twitter summary cards for your blogdown post
}
\examples{
add_twitter_card()
}
