% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_twitter_card.R
\name{add_twitter_card}
\alias{add_twitter_card}
\title{Create twitter summary cards}
\usage{
add_twitter_card(card_type = c("summary", "summary_large_image"), title,
  description = NULL, user = NULL, image = NULL, image_alt = NULL,
  file = "twitter_card.html")
}
\arguments{
\item{card_type}{Required. Twitter card type. Must be "summary" or "summary_large_image".}

\item{title}{Required. Title for card content. Max 70 characters.}

\item{description}{Description for card content, max 200 characters.}

\item{user}{Twitter user id for the card to be linked to.}

\item{image}{Link for the content image. Must be absolute path, i.e. http}

\item{image_alt}{Alternate text incase image does not load.}

\item{file}{Required. File name to put in the header of Rmd doc, must be .html.}
}
\description{
An easy way to create twitter summary cards for your blogdown post.
    For more details on the meta tags see https://developer.twitter.com/en/docs/tweets/optimize-with-cards/overview/summary.html
}
\examples{
add_twitter_card(title = "HeadR is great!", file = "hello.html")
}
