% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{duplicates}
\alias{duplicates}
\title{Count and tag duplicates (in terms of a variable list)}
\usage{
duplicates(DT, by.vars, tag = FALSE)
}
\arguments{
\item{DT}{a data.table}

\item{by.vars}{a list of variables}

\item{tag}{should the fn create an indicator for duplicate obs?}
}
\description{

}
\details{
duplicates() resembles Stata command of same name to count and tag repeat observations
Warning: duplicates() uses assign by reference in data.table
You must library(HeadR) to use this function
Revision note: use copy() to avoid changes to original DT
}
\examples{
EG <- data.table(x=LETTERS[1:4],y=LETTERS[1:2],z=1:2)
duplicates(EG,by.vars=c("x","y","z"))
duplicates(EG,by.vars=c("y","z"))
EG <- duplicates(EG,by.vars=c("y","z"),tag=TRUE)
}
