% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ni.R
\name{\%ni\%}
\alias{\%ni\%}
\title{not-in operator to exclude elements of a set.}
\usage{
x \%ni\% y
}
\arguments{
\item{x}{a vector, usually a set of id strings}
}
\description{

}
\details{
Negates the \%in\% operator. Returns a logical. Equivalent to keep if x != y\link{1} and x!= y\link{2}, etc.
}
