% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_str.R
\name{\%+\%}
\alias{\%+\%}
\title{paste two strings into one with no space between them
#' @description}
\usage{
x \%+\% y
}
\arguments{
\item{x}{a vector, usually string}

\item{y}{a vector, usually string}
}
\description{
paste two strings into one with no space between them
#' @description
}
\details{
an operator to do a simple paste0
}
\examples{
n = 5
"filename" \%+\% n \%+\% ".pdf"
c(1,NA,3,NA) \%+\% c(7,5,NA,NA)
}
