% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipolate.R
\name{ipolate}
\alias{ipolate}
\title{Linear interpolation of missing data}
\usage{
ipolate(xvar, yvar)
}
\arguments{
\item{xvar}{the variable to interpolate using (usually year)}

\item{yvar}{the name of the variable with missing data}
}
\description{
Linear interpolation of missing data
}
\details{
ipolate resembles Stata command of same name, very little testing on this fn
}
