% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice.R
\name{nice}
\alias{nice}
\title{Make numbers look nice for a table}
\usage{
nice(x, digits = 2)
}
\arguments{
\item{x}{a number}

\item{digits}{the number of digits to show to right of decimal point}
}
\description{
Warning the result is not numeric anymore, so do not do any calculations
}
\examples{
 nice(c(17.0003,4.726677))
nice(c(17.0003,4.726677),digits=0)
}
