% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_left.R
\name{st_left}
\alias{st_left}
\title{Create a new string of text to the left of a "_"}
\usage{
st_left(x, sep = "_")
}
\arguments{
\item{x}{a character vector}

\item{sep}{a separator character (defaults to "_", don't use backslash)}
}
\description{
Create a new string of text to the left of a "_"
}
\details{
use to separate a string into just the part to left of a delimiter such as "-" or "_", also see st_right
}
\examples{
st_left(c("WUHU_1","WUHU_2","WU_HU_WU","WUHU"))
st_left(c("WUHU_1","WUHU/2","WU_HU/WU","WUHU"),"/")
}
