% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texout.R
\name{texout}
\alias{texout}
\title{create a column in a data.table with tabular info}
\usage{
texout(x, ndigits = 2)
}
\arguments{
\item{x}{a list of variables
ndigits number of digits after decimal point}
}
\description{

}
\examples{
mydt <- data.table(ida=LETTERS[1:10],idn=1:10,x=runif(10),y=runif(10))
mydt[,output := texout(list(ida,idn,x,y))]
writeLines(mydt$output)
mydt[,output := texout(list(ida,idn,x,y),3)]
writeLines(mydt$output)
}
